fix(module): revert tailwind config from #1272
